Split out tepmo synthesis from detection
Make TempoAnalyserComponent only deduce the tempo from the samples it
is given and introduce a component, TempoSynthesizerComponent, that
will handle synthesising finer grained control of the tempo send to
visuals system from the raw detected tempo and beats.

Currently only part of the work has been done to extract out the
synthesis from the detection.